Send bitrate value to MM on basis of MTU and sample rate.
When MM does A2DP_CTRL_GET_CODEC_CONFIG BT HOST need to send the
MTU and bitrate which would be calculated based on MTU and sample rate.
So that DSP would encode the Audio data on basis of these parameters,
so that decoding of Audio data at remote end would be proper and streaming
would be smooth.

CRs-Fixed: 2402346
Change-Id: I6c964311dfa5db335d3b9dd21c60d1f9dcc61ebe