Use new RPA only when LE set random address cmd is sent
BT stack computes the new RPA every 15 mins for its
own address and uses this RPA even when LE set random
address command is not successfully sent to controller.

LE set random address command is not sent to the
controller by BT stack when LE direct connection is
pending. But the computed RPA is used by BT stack for
DH key check calculation during LE pairing.
Hence, there is a mismatch between the DHKey check
values computed on the local and the peer devices
and the pairing fails.

This change makes sure that the BT stack uses the
computed RPA only when LE set random address command
is successfully sent to the controller.

Change-Id: I5ea4d4716b9ccc6bc5c153577b9ba2ac48c5db18