A2DP SINK: fix regression issue of no voice of all codecs
The index of current playing device is set to btif_max_av_clients
if no device is active, the invalid index will cause sho is not
required, and the current playing index will not be updated and sink
media data will not be enqueued. Modify the invalid index in
btif_is_a2dp_sink_handoff_required to fix the issue.

CRs-Fixed: 2479867

Change-Id: Ib37248ddeb96717d07ffa33ab2940bfed2e1d878
Signed-off-by: Liuliu Zhao <liulzhao@codeaurora.org>